requirements.txt
anyio==3.7.1
bcrypt==4.0.1
beanie==1.20.0
click==8.1.4
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
httptools==0.6.0
idna==3.4
lazy-model==0.0.5
motor==3.2.0
passlib==1.7.4
pydantic==1.10.11
pymongo==4.4.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.7.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
